+++ ruoyi-business/src/main/java/com/ruoyi/business/masterdata/domain/MasterDataRecord.java
@@ -63,7 +63,6 @@ public class MasterDataRecord extends BaseEntity
         this.resource = resource;
     }
 
-    @NotBlank(message = "编码不能为空")
     public String getItemCode()
     {
         return itemCode;